React Basics 15 - Parent Child Relationship (Composite Design Pattern) & Lifting State Up
